renamed 'ticket-api/test/' package name, checked README.md, added absolute path for kafka data volume
